return jwt token on signin
