+++ app.js
@@ -5,6 +5,7 @@ const passport = require('passport');
 const morgan = require('morgan');
 const cors = require('cors');
 const routes = require('./routes');
+const middlewares = require('./middlewares');
 // const { logger } = require('./utils');
 const { NODE_ENV } = process.env;
 
@@ -22,6 +23,8 @@ app.use(passport.initialize());
 app.use(methodOverride('X-HTTP-Method-Override'));
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
+
+Object.values(middlewares).forEach((middleware) => app.use(middleware));
 
 app.use('/api', routes);
