+++ middlewares/index.js
@@ -1,0 +1,5 @@
+const jwtParser = require('./jwtParser');
+
+module.exports = {
+  jwtParser,
+};
